pyproject.toml
[project]
name = "mcpdemolangchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "duckduckgo-search>=8.0.4",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
